Move FileSize and FileSizeFormatProvider into SystemUtil folder of CommonUtil project
