+++ pwiz_tools/Skyline/Util/UtilUIExtra.cs
@@ -1,4 +1,22 @@
-﻿using System;
+﻿/*
+ * Original author: Brendan MacLean <brendanx .at. u.washington.edu>,
+ *                  MacCoss Lab, Department of Genome Sciences, UW
+ *
+ * Copyright 2009 University of Washington - Seattle, WA
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+using System;
 using System.Diagnostics;
 using System.Drawing;
 using System.Runtime.InteropServices;
@@ -255,86 +273,6 @@ EndSelection:<<<<<<<3
         #endregion // Write to Clipboard
     }
 
-    
-    /// <summary>
-    /// String formatting for file size.
-    /// Downloaded from: http://flimflan.com/blog/FileSizeFormatProvider.aspx 
-    /// </summary>
-    public class FileSizeFormatProvider : IFormatProvider, ICustomFormatter
-    {
-        public object GetFormat(Type formatType)
-        {
-            if (formatType == typeof(ICustomFormatter)) return this;
-            return null;
-        }
-
-        private const string FILE_SIZE_FORMAT = "fs";
-        private const Decimal ONE_KILO_BYTE = 1024M;
-        private const Decimal ONE_MEGA_BYTE = ONE_KILO_BYTE * 1024M;
-        private const Decimal ONE_GIGA_BYTE = ONE_MEGA_BYTE * 1024M;
-
-        public string Format(string format, object arg, IFormatProvider formatProvider)
-        {
-            if (format == null || !format.StartsWith(FILE_SIZE_FORMAT))
-            {
-                return DefaultFormat(format, arg, formatProvider);
-            }
-
-            if (arg is string)
-            {
-                return DefaultFormat(format, arg, formatProvider);
-            }
-
-            Decimal size;
-
-            try
-            {
-                size = Convert.ToDecimal(arg);
-            }
-            catch (InvalidCastException)
-            {
-                return DefaultFormat(format, arg, formatProvider);
-            }
-
-            string suffix;
-
-            if (size > ONE_GIGA_BYTE)
-            {
-                size /= ONE_GIGA_BYTE;
-                suffix = @" GB";
-            }
-            else if (size > ONE_MEGA_BYTE)
-            {
-                size /= ONE_MEGA_BYTE;
-                suffix = @" MB";
-            }
-            else if (size > ONE_KILO_BYTE)
-            {
-                size /= ONE_KILO_BYTE;
-                suffix = @" KB";
-            }
-            else
-            {
-                suffix = @" B";
-            }
-
-            string precision = format.Substring(2);
-            if (String.IsNullOrEmpty(precision))
-                precision = @"2";
-            string formatString = @"{0:N" + precision + @"}{1}";  // Avoid ReSharper analysis
-            return String.Format(formatString, size, suffix);
-        }
-
-        private static string DefaultFormat(string format, object arg, IFormatProvider formatProvider)
-        {
-            IFormattable formattableArg = arg as IFormattable;
-            if (formattableArg != null)
-            {
-                return formattableArg.ToString(format, formatProvider);
-            }
-            return arg.ToString();
-        }
-    }
 
     public static class ComboHelper
     {
